break out oras.go into separate files
