+++ src/pkg/oci/transport.go
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2021-Present The Zarf Authors
 
-// Package utils provides generic helper functions.
-package utils
+// Package oci contains functions for interacting with Zarf packages stored in OCI registries.
+package oci
 
 import (
 	"io"
