+++ src/test/e2e/51_oci_compose_test.go
@@ -25,7 +25,6 @@ import (
 type SkeletonSuite struct {
 	suite.Suite
 	*require.Assertions
-	Remote    *utils.OrasRemote
 	Reference registry.Reference
 }
